feat: add debug mode
Now it is possible to start gorched with `--debug` flag and turn on
debug mode.

In debug mode it is possible to collect log messages with
object `Debug` and show them on screen. Current FPS is shown also in
debug mode. Debug messages can be hidden and shown again with `Ctrl+D`.