fix: cut terrain around tank
When tank sprite would be draw over part of terrain this part of terrain
is removed.

There was need to change how terrain is drawn to keep terrain texture
unmodified by cuts. Terrain is now pre-rendered into it's columns and
these coloumns are drawn at the end. This would be useful later for
terrain modifications after bullet hit.