refactor: represent world with custom extension of termloop.BaseLevel
Custom level implementation allows change how entities are removed. Now when RemoveEnity is called entity is only registered to array and it will be removed on next Tick. This enables removing entities from Draw method without wierd glitches caused by modifing Entities array in the middle of for loop.

For the future there is potential for easier access to some entities or
to provide custom parameters to Draw and Tick methods.