feat: add terrain destruction
Now there are holes in the terrain after bullet hit the ground. Radius of
the hole depends on the `strength` attribute of bullet which is for now
hardcoded to 3.

There is no gravity yet, so terrain parts, trees and tanks stays in the
air even there is hole under.

Terrain representation was slightly modified. Now terrain columns can be
split to multiple smaller columns and therefor there can be zero or more
terrain columns for each x coordinate.