refactor: move message boxes to the screen and out of controls
There was need to draw message boxes on the screen always on top of
other level entities. Therefor new entity named HUD was created which
is added to the screen instead of to the level. HUD holds reference to
currenlty active message box.

Controls logic for showing / hiding message boxes was simplified to
calling few methods on HUD.

Message box creating was moved out from controls and utils to separate
file.