feat: add explosion of the bullet
When bullet hits the ground or the tank there is an explosion now.

Explosion size is based on the strength of bullet. It "lives" for around
one second. It can create hole into terrain with same radius as it's strength.

Explosion cannot kill the tank yet. Tank still needs to be hit by
bullet.