refactor: extract terrain destruction logic to separate entity
Terrain cutting logic was quite messy and there were some bugs when
multiple explosions were colliding with terrain in the same time.

New entity terrain.Cutter was created which takes care about all
terrain cuttings. It is added to the world with terrain columns and it
is regularly processing any cuts in it's Draw methods.

Beside that terrain was moved to separate package and splitted to
multiple files.

Similarly debug and physics were moved to separate packages to
avoid cyclic dependencies.