fix: improve initial bullet position
Initial position of bullet was updated to look more like coming from
cannon.

Initial position has effect also on terrain collisions when tank is close to
the terrain. E.g. for angles bigger then 105 degrees there was collision not
with closest but with second closest terrain column because of bad
initial bullet position.