feat: add helath to tanks and make explosions to take damage
Now each tank has health set to 100 after new round start. When there is
an explosion near the tank it can take some damage to this tank. Amount
of taken damage depends on the distance of explosion center from the
tank and on the explosion's strength.

When tank's health goes on zero it will explode.