refactor: create entity for round management
Now there is separate entity with logic around rounds. It manages
round state and it provides functionality for transitions to next rounds.

Previously this fucntionality was on multiple places in Game and in
Controls.

Additionally there was fix for hadnling tank killed by explosion. It was
not recognized before, round was not finished and socore of shooter was
not updated.