refactor: add simple ui support
Support for creating simple forms with some basic components was added.
These forms should be used for implementaion of menu.

Main entry point is Form, which can be added to the termloop.Screen.
It is responsible for drawing other components and for managing input
events.

There is support for creating hierarchies of components using Container.

Currently provided components:

 - Text - shows some text
 - Value - shows some numeric value
 - Button - can be pressed to invoke some action

There is also support for showing simple form with message in
MessageBox.

There is also support for creating container from formatted text using
NewFromatPane function.