refactor: move hud to separate package and use new ui support
After new ui components support was introduced it was used for info and
socre message boxes.

As they belong to HUD they were moved to the same package with HUD.

To be avoid circular dependencies there was introduced new package
called core. This package now contains only Player definition
but in the future it should contain core game logic independned of game engine.